requirements.txt
# Streamlit Frontend Requirements
streamlit>=1.28.0
requests>=2.31.0
pandas>=1.5.0
plotly>=5.15.0
streamlit-option-menu>=0.3.6
streamlit-lottie>=0.0.5
streamlit-aggrid>=0.3.4
pillow>=9.5.0